@@ -35,9 +35,9 @@
     "promise": "8.0.1",
     "prop-types": "^15.6.0",
     "r": "0.0.5",
-    "react": "^16.0.0",
+    "react": "^16.14.0",
     "react-dev-utils": "^4.0.1",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.14.0",
     "react-redux": "^5.1.2",
     "react-router-dom": "^4.2.2",
     "react-test-renderer": "^17.0.2",
