@@ -1,16 +1,17 @@
-import React, { Component } from "react";
+import React, { useEffect } from "react";
 import { connect } from "react-redux";
 import { Redirect } from "react-router-dom";
 import * as actions from "../../../store/actions/index";
 
-class Logout extends Component {
-	componentDidMount() {
-		this.props.onLogout();
-	}
-	render() {
-		return <Redirect to="/" />;
-	}
-}
+const logout = (props) => {
+
+	const {onLogout } = props;
+
+	useEffect(() => {
+		onLogout();
+	}, [onLogout]);
+	return <Redirect to="/" />;
+};
 
 const mapDispatchToProps = (dispatch) => {
 	return {
@@ -18,4 +19,4 @@ const mapDispatchToProps = (dispatch) => {
 	};
 };
 
-export default connect(null, mapDispatchToProps)(Logout);
+export default connect(null, mapDispatchToProps)(logout);
