@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useEffect, Suspense } from "react";
 import { Route, Switch, withRouter, Redirect } from "react-router-dom";
 import { connect } from "react-redux";
 import * as actions from "./store/actions/index";
@@ -6,54 +6,55 @@ import * as actions from "./store/actions/index";
 import Layout from "./hoc/Layout/Layout";
 import BurgerBuilder from "./containers/BurgerBuilder/BurgerBuilder";
 import Logout from "./containers/Auth/Logout/Logout";
-import asyncComponent from "./asyncComponent/asyncComponent";
 
-const asyncCheckout = asyncComponent(() => {
+const Checkout = React.lazy(() => {
 	return import("./containers/Checkout/Checkout");
 });
 
-const asyncOrders = asyncComponent(() => {
+const Orders = React.lazy(() => {
 	return import("./containers/Orders/Orders");
 });
 
-const asyncAuth = asyncComponent(() => {
+const Auth = React.lazy(() => {
 	return import("./containers/Auth/Auth");
 });
 
-class App extends Component {
-	componentDidMount() {
-		this.props.onAuthCheck();
-	}
+const app = (props) => {
+	const { onAuthCheck } = props;
 
-	render() {
-		let routes = (
+	useEffect(() => {
+		onAuthCheck();
+	}, [onAuthCheck]);
+
+	let routes = (
+		<Switch>
+			<Route path="/auth" render={(props) => <Auth {...props} />} />
+			<Route path="/" exact component={BurgerBuilder} />
+			<Redirect to="/" />
+		</Switch>
+	);
+
+	if (props.isAuthenticated) {
+		routes = (
 			<Switch>
-				<Route path="/auth" component={asyncAuth} />
+				<Route path="/checkout" render={(props) => <Checkout {...props} />} />
+				<Route path="/orders" render={(props) => <Orders {...props} />} />
+				<Route path="/logout" component={Logout} />
+				<Route path="/auth" render={(props) => <Auth {...props} />} />
 				<Route path="/" exact component={BurgerBuilder} />
 				<Redirect to="/" />
 			</Switch>
 		);
-
-		if (this.props.isAuthenticated) {
-			routes = (
-				<Switch>
-					<Route path="/checkout" component={asyncCheckout} />
-					<Route path="/orders" component={asyncOrders} />
-					<Route path="/logout" component={Logout} />
-					<Route path="/auth" component={asyncAuth} />
-					<Route path="/" exact component={BurgerBuilder} />
-					<Redirect to="/" />
-				</Switch>
-			);
-		}
-
-		return (
-			<div>
-				<Layout>{routes}</Layout>
-			</div>
-		);
 	}
-}
+
+	return (
+		<div>
+			<Layout>
+				<Suspense fallback={<p>Loading...</p>}>{routes}</Suspense>
+			</Layout>
+		</div>
+	);
+};
 
 const mapStateToProps = (state) => {
 	return {
@@ -67,4 +68,4 @@ const mapDispatchToProps = (dispatch) => {
 	};
 };
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(App));
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(app));
