@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState, useEffect } from "react";
 
 import { Redirect } from "react-router-dom";
 import Input from "../../components/UI/Input/Input";
@@ -9,128 +9,116 @@ import { connect } from "react-redux";
 import Spinner from "../../components/UI/Spinner/Spinner";
 import { updateObject, checkValidity } from "../../shared/utility";
 
-class Auth extends Component {
-	state = {
-		controls: {
-			email: {
-				elementType: "input",
-				elementConfig: {
-					type: "email",
-					placeholder: "E-mail Address",
-				},
-				value: "",
-				validation: {
-					required: true,
-					isEmail: true,
-				},
-				valid: false,
-				touched: false,
+const auth = (props) => {
+	const [controls, setControls] = useState({
+		email: {
+			elementType: "input",
+			elementConfig: {
+				type: "email",
+				placeholder: "E-mail Address",
 			},
-			password: {
-				elementType: "input",
-				elementConfig: {
-					type: "password",
-					placeholder: "Password",
-				},
-				value: "",
-				validation: {
-					required: true,
-					minLength: 6,
-				},
-				valid: false,
-				touched: false,
+			value: "",
+			validation: {
+				required: true,
+				isEmail: true,
 			},
+			valid: false,
+			touched: false,
 		},
-		isSignUp: true,
-	};
+		password: {
+			elementType: "input",
+			elementConfig: {
+				type: "password",
+				placeholder: "Password",
+			},
+			value: "",
+			validation: {
+				required: true,
+				minLength: 6,
+			},
+			valid: false,
+			touched: false,
+		},
+	});
 
-	componentDidMount() {
-		if (
-			this.props.redirectionPath === "/checkout" &&
-			!this.props.burgerBuilding
-		) {
-			this.props.onSetAuthRedirectionPath();
+	const [isSignUp, setIsSignUp] = useState(true);
+
+	const { redirectionPath, burgerBuilding, onSetAuthRedirectionPath } = props;
+
+	useEffect(() => {
+		if (redirectionPath === "/checkout" && !burgerBuilding) {
+			onSetAuthRedirectionPath();
 		}
-	}
+	}, [redirectionPath, burgerBuilding, onSetAuthRedirectionPath]);
 
-	inputChangedHandler = (event, controlName) => {
-		const updatedControls = updateObject(this.state.controls, {
-			[controlName]: updateObject(this.state.controls[controlName], {
+	const inputChangedHandler = (event, controlName) => {
+		const updatedControls = updateObject(controls, {
+			[controlName]: updateObject(controls[controlName], {
 				value: event.target.value,
 				valid: checkValidity(
 					event.target.value,
-					this.state.controls[controlName].validation
+					controls[controlName].validation
 				),
 				touched: true,
 			}),
 		});
-		this.setState({ controls: updatedControls });
+		setControls(updatedControls);
 	};
 
-	submitHandler = (event) => {
+	const submitHandler = (event) => {
 		event.preventDefault();
-		this.props.onAuth(
-			this.state.controls.email.value,
-			this.state.controls.password.value,
-			this.state.isSignUp
-		);
+		props.onAuth(controls.email.value, controls.password.value, isSignUp);
 	};
 
-	switchFrom = () => {
-		this.setState((pervState) => {
-			return {
-				isSignUp: !pervState.isSignUp,
-			};
+	const switchFrom = () => {
+		setIsSignUp(!isSignUp);
+	};
+
+	const formElementsArray = [];
+	for (let key in controls) {
+		formElementsArray.push({
+			id: key,
+			config: controls[key],
 		});
-	};
-
-	render() {
-		const formElementsArray = [];
-		for (let key in this.state.controls) {
-			formElementsArray.push({
-				id: key,
-				config: this.state.controls[key],
-			});
-		}
-
-		let form = formElementsArray.map((el) => (
-			<Input
-				key={el.id}
-				elementType={el.config.elementType}
-				elementConfig={el.config.elementConfig}
-				value={el.config.value}
-				invalid={!el.config.valid}
-				shouldValidate={el.config.validation}
-				touched={el.config.touched}
-				changed={(event) => this.inputChangedHandler(event, el.id)}
-			/>
-		));
-
-		if (this.props.loading) {
-			form = <Spinner />;
-		}
-
-		let authRedirect = null;
-		if (this.props.isAuthenticated) {
-			authRedirect = <Redirect to={this.props.redirectionPath} />;
-		}
-		return (
-			<div className={classes.Auth}>
-				<form onSubmit={this.submitHandler}>
-					{form}
-					<Button btnType="Success">Submit</Button>
-				</form>
-				{this.props.error ? (
-					<p style={{ color: "red" }}>{this.props.error.message}</p>
-				) : null}
-				<Button clicked={this.switchFrom} btnType="Danger">
-					switch to {this.state.isSignUp ? "Sign in" : "Sign up"}
-				</Button>
-				{authRedirect}
-			</div>
-		);
 	}
-}
+
+	let form = formElementsArray.map((el) => (
+		<Input
+			key={el.id}
+			elementType={el.config.elementType}
+			elementConfig={el.config.elementConfig}
+			value={el.config.value}
+			invalid={!el.config.valid}
+			shouldValidate={el.config.validation}
+			touched={el.config.touched}
+			changed={(event) => inputChangedHandler(event, el.id)}
+		/>
+	));
+
+	if (props.loading) {
+		form = <Spinner />;
+	}
+
+	let authRedirect = null;
+	if (props.isAuthenticated) {
+		authRedirect = <Redirect to={props.redirectionPath} />;
+	}
+	return (
+		<div className={classes.Auth}>
+			<form onSubmit={submitHandler}>
+				{form}
+				<Button btnType="Success">Submit</Button>
+			</form>
+			{props.error ? (
+				<p style={{ color: "red" }}>{props.error.message}</p>
+			) : null}
+			<Button clicked={switchFrom} btnType="Danger">
+				switch to {isSignUp ? "Sign in" : "Sign up"}
+			</Button>
+			{authRedirect}
+		</div>
+	);
+};
 
 const mapStateToProps = (state) => {
 	return {
@@ -151,4 +139,4 @@ const mapDispatchToProps = (dispatch) => {
 	};
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Auth);
+export default connect(mapStateToProps, mapDispatchToProps)(auth);
