@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState, useEffect } from "react";
 import { connect } from "react-redux";
 
 import Aux from "../../hoc/Auxillary/Auxillary";
@@ -11,16 +11,17 @@ import withErrorHandler from "../../hoc/withErrorHandler/withErrorHandler";
 import * as actions from "../../store/actions/index";
 import axios from "../../axios-orders";
 
-export class BurgerBuilder extends Component {
-	state = {
-		purchasing: false,
-	};
+const burgerBuilder = props => {
 
-	componentDidMount() {
-		this.props.onInitIngredients();
-	}
+	const [purchasing, setPurchasing] = useState(false)
+	const {onInitIngredients} = props;
 
-	updatePurchaseState(ingredients) {
+
+	useEffect(() =>  {
+		onInitIngredients();
+	}, [onInitIngredients])
+
+	const updatePurchaseState = (ingredients) => {
 		const sum = Object.keys(ingredients)
 			.map((igKey) => {
 				return ingredients[igKey];
@@ -31,73 +32,71 @@ export class BurgerBuilder extends Component {
 		return sum > 0;
 	}
 
-	purchaseHandler = () => {
-		if (this.props.isAuthenticated) {
-			this.setState({ purchasing: true });
+	const purchaseHandler = () => {
+		if (props.isAuthenticated) {
+			setPurchasing(true);
 		} else {
-			this.props.onSetAuthRedirectionPath("/checkout");
-			this.props.history.push("/auth");
+			props.onSetAuthRedirectionPath("/checkout");
+			props.history.push("/auth");
 		}
 	};
 
-	purchaseCancelHandler = () => {
-		this.setState({ purchasing: false });
+	const purchaseCancelHandler = () => {
+		setPurchasing(false);
 	};
 
-	purchaseContinueHandler = () => {
-		this.props.onInitPurchase();
-		this.props.history.push("/checkout");
+	const purchaseContinueHandler = () => {
+		props.onInitPurchase();
+		props.history.push("/checkout");
 	};
 
-	render() {
 		const disabledInfo = {
-			...this.props.ings,
+			...props.ings,
 		};
 		for (let key in disabledInfo) {
 			disabledInfo[key] = disabledInfo[key] <= 0;
 		}
 		let orderSummary = null;
-		let burger = this.props.error ? (
+		let burger = props.error ? (
 			<p>Ingredients can't be loaded!</p>
 		) : (
 			<Spinner />
 		);
 
-		if (this.props.ings) {
+		if (props.ings) {
 			burger = (
 				<Aux>
-					<Burger ingredients={this.props.ings} />
+					<Burger ingredients={props.ings} />
 					<BuildControls
-						ingredientAdded={this.props.onIngredientAdded}
-						ingredientRemoved={this.props.onIngredientRemoved}
+						ingredientAdded={props.onIngredientAdded}
+						ingredientRemoved={props.onIngredientRemoved}
 						disabled={disabledInfo}
-						purchasable={this.updatePurchaseState(this.props.ings)}
-						ordered={this.purchaseHandler}
-						isAuth={this.props.isAuthenticated}
-						price={this.props.price}
+						purchasable={updatePurchaseState(props.ings)}
+						ordered={purchaseHandler}
+						isAuth={props.isAuthenticated}
+						price={props.price}
 					/>
 				</Aux>
 			);
 			orderSummary = (
 				<OrderSummary
-					ingredients={this.props.ings}
-					price={this.props.price}
-					purchaseCancelled={this.purchaseCancelHandler}
-					purchaseContinued={this.purchaseContinueHandler}
+					ingredients={props.ings}
+					price={props.price}
+					purchaseCancelled={purchaseCancelHandler}
+					purchaseContinued={purchaseContinueHandler}
 				/>
 			);
 		}
 		return (
 			<Aux>
 				<Modal
-					show={this.state.purchasing}
-					modalClosed={this.purchaseCancelHandler}>
+					show={purchasing}
+					modalClosed={purchaseCancelHandler}>
 					{orderSummary}
 				</Modal>
 				{burger}
 			</Aux>
 		);
-	}
 }
 
 const mapStateToProps = (state) => {
@@ -124,4 +123,4 @@ const mapDispatchToProps = (dispatch) => {
 export default connect(
 	mapStateToProps,
 	mapDispatchToProps
-)(withErrorHandler(BurgerBuilder, axios));
+)(withErrorHandler(burgerBuilder, axios));
