@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import { connect } from "react-redux";
 
 import Button from "../../../components/UI/Button/Button";
@@ -10,126 +10,121 @@ import withErrorHandler from "../../../hoc/withErrorHandler/withErrorHandler";
 import * as actions from "../../../store/actions/index";
 import { updateObject, checkValidity } from "../../../shared/utility";
 
-class ContactData extends Component {
-	state = {
-		orderForm: {
-			name: {
-				elementType: "input",
-				elementConfig: {
-					type: "text",
-					placeholder: "Your Name",
-				},
-				value: "",
-				validation: {
-					required: true,
-				},
-				valid: false,
-				touched: false,
+const contactData = props => {
+	const [orderForm, setOrderForm] = useState({
+		name: {
+			elementType: "input",
+			elementConfig: {
+				type: "text",
+				placeholder: "Your Name",
 			},
-			street: {
-				elementType: "input",
-				elementConfig: {
-					type: "text",
-					placeholder: "Street",
-				},
-				value: "",
-				validation: {
-					required: true,
-				},
-				valid: false,
-				touched: false,
+			value: "",
+			validation: {
+				required: true,
 			},
-			zipCode: {
-				elementType: "input",
-				elementConfig: {
-					type: "text",
-					placeholder: "ZIP Code",
-				},
-				value: "",
-				validation: {
-					required: true,
-					minLength: 5,
-					maxLength: 5,
-					isNumeric: true,
-				},
-				valid: false,
-				touched: false,
-			},
-			country: {
-				elementType: "input",
-				elementConfig: {
-					type: "text",
-					placeholder: "Country",
-				},
-				value: "",
-				validation: {
-					required: true,
-				},
-				valid: false,
-				touched: false,
-			},
-			email: {
-				elementType: "input",
-				elementConfig: {
-					type: "email",
-					placeholder: "Your E-Mail",
-				},
-				value: "",
-				validation: {
-					required: true,
-					isEmail: true,
-				},
-				valid: false,
-				touched: false,
-			},
-			deliveryMethod: {
-				elementType: "select",
-				elementConfig: {
-					options: [
-						{ value: "fastest", displayValue: "Fastest" },
-						{ value: "cheapest", displayValue: "Cheapest" },
-					],
-				},
-				value: "fastest",
-				validation: {},
-				valid: true,
-			},
+			valid: false,
+			touched: false,
 		},
-		formIsValid: false,
-	};
+		street: {
+			elementType: "input",
+			elementConfig: {
+				type: "text",
+				placeholder: "Street",
+			},
+			value: "",
+			validation: {
+				required: true,
+			},
+			valid: false,
+			touched: false,
+		},
+		zipCode: {
+			elementType: "input",
+			elementConfig: {
+				type: "text",
+				placeholder: "ZIP Code",
+			},
+			value: "",
+			validation: {
+				required: true,
+				minLength: 5,
+				maxLength: 5,
+				isNumeric: true,
+			},
+			valid: false,
+			touched: false,
+		},
+		country: {
+			elementType: "input",
+			elementConfig: {
+				type: "text",
+				placeholder: "Country",
+			},
+			value: "",
+			validation: {
+				required: true,
+			},
+			valid: false,
+			touched: false,
+		},
+		email: {
+			elementType: "input",
+			elementConfig: {
+				type: "email",
+				placeholder: "Your E-Mail",
+			},
+			value: "",
+			validation: {
+				required: true,
+				isEmail: true,
+			},
+			valid: false,
+			touched: false,
+		},
+		deliveryMethod: {
+			elementType: "select",
+			elementConfig: {
+				options: [
+					{ value: "fastest", displayValue: "Fastest" },
+					{ value: "cheapest", displayValue: "Cheapest" },
+				],
+			},
+			value: "fastest",
+			validation: {},
+			valid: true,
+		},
+	});
+	const [formIsValid, setFormIsValid] = useState(false);
 
-	orderHandler = (event) => {
+	const orderHandler = (event) => {
 		event.preventDefault();
 
 		const formData = {};
-		for (let formElementIdentifier in this.state.orderForm) {
+		for (let formElementIdentifier in orderForm) {
 			formData[formElementIdentifier] =
-				this.state.orderForm[formElementIdentifier].value;
+				orderForm[formElementIdentifier].value;
 		}
 		const order = {
-			ingredients: this.props.ings,
-			price: this.props.price,
+			ingredients: props.ings,
+			price: props.price,
 			orderData: formData,
-			userId: this.props.userId,
+			userId: props.userId,
 		};
 
-		this.props.onOrderBurger(order, this.props.token);
+		props.onOrderBurger(order, props.token);
 	};
 
-	inputChangedHandler = (event, inputIdentifier) => {
-		const updatedFormElement = updateObject(
-			this.state.orderForm[inputIdentifier],
-			{
-				value: event.target.value,
-				valid: checkValidity(
-					event.target.value,
-					this.state.orderForm[inputIdentifier].validation
-				),
-				touched: true,
-			}
-		);
+	const inputChangedHandler = (event, inputIdentifier) => {
+		const updatedFormElement = updateObject(orderForm[inputIdentifier], {
+			value: event.target.value,
+			valid: checkValidity(
+				event.target.value,
+				orderForm[inputIdentifier].validation
+			),
+			touched: true,
+		});
 
-		const updatedOrderForm = updateObject(this.state.orderForm, {
+		const updatedOrderForm = updateObject(orderForm, {
 			[inputIdentifier]: updatedFormElement,
 		});
 
@@ -137,47 +132,46 @@ class ContactData extends Component {
 		for (let inputIdentifier in updatedOrderForm) {
 			formIsValid = updatedOrderForm[inputIdentifier].valid && formIsValid;
 		}
-		this.setState({ orderForm: updatedOrderForm, formIsValid: formIsValid });
+		setOrderForm(updatedOrderForm);
+		setFormIsValid(formIsValid);
 	};
 
-	render() {
-		const formElementsArray = [];
-		for (let key in this.state.orderForm) {
-			formElementsArray.push({
-				id: key,
-				config: this.state.orderForm[key],
-			});
-		}
-		let form = (
-			<form onSubmit={this.orderHandler}>
-				{formElementsArray.map((formElement) => (
-					<Input
-						key={formElement.id}
-						elementType={formElement.config.elementType}
-						elementConfig={formElement.config.elementConfig}
-						value={formElement.config.value}
-						invalid={!formElement.config.valid}
-						shouldValidate={formElement.config.validation}
-						touched={formElement.config.touched}
-						changed={(event) => this.inputChangedHandler(event, formElement.id)}
-					/>
-				))}
-				<Button btnType="Success" disabled={!this.state.formIsValid}>
-					ORDER
-				</Button>
-			</form>
-		);
-		if (this.props.loading) {
-			form = <Spinner />;
-		}
-		return (
-			<div className={classes.ContactData}>
-				<h4>Enter your Contact Data</h4>
-				{form}
-			</div>
-		);
+	const formElementsArray = [];
+	for (let key in orderForm) {
+		formElementsArray.push({
+			id: key,
+			config: orderForm[key],
+		});
 	}
-}
+	let form = (
+		<form onSubmit={orderHandler}>
+			{formElementsArray.map((formElement) => (
+				<Input
+					key={formElement.id}
+					elementType={formElement.config.elementType}
+					elementConfig={formElement.config.elementConfig}
+					value={formElement.config.value}
+					invalid={!formElement.config.valid}
+					shouldValidate={formElement.config.validation}
+					touched={formElement.config.touched}
+					changed={(event) => inputChangedHandler(event, formElement.id)}
+				/>
+			))}
+			<Button btnType="Success" disabled={!formIsValid}>
+				ORDER
+			</Button>
+		</form>
+	);
+	if (props.loading) {
+		form = <Spinner />;
+	}
+	return (
+		<div className={classes.ContactData}>
+			<h4>Enter your Contact Data</h4>
+			{form}
+		</div>
+	);
+};
 
 const mapStateToProps = (state) => {
 	return {
@@ -199,4 +193,4 @@ const mapDispatchToProps = (dispatch) => {
 export default connect(
 	mapStateToProps,
 	mapDispatchToProps
-)(withErrorHandler(ContactData, axios));
+)(withErrorHandler(contactData, axios));
