@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useEffect } from "react";
 import { connect } from "react-redux";
 
 import Order from "../../components/Order/Order";
@@ -7,25 +7,26 @@ import withErrorHandler from "../../hoc/withErrorHandler/withErrorHandler";
 import * as actions from "../../store/actions/index";
 import Spinner from "../../components/UI/Spinner/Spinner";
 
-class Orders extends Component {
-	componentDidMount() {
-		this.props.onFetchOrders(this.props.token, this.props.userId);
-	}
+const orders = (props) => {
 
-	render() {
-		let orders = <Spinner />;
-		if (!this.props.loading) {
-			orders = this.props.orders.map((order) => (
-				<Order
-					key={order.id}
-					ingredients={order.ingredients}
-					price={order.price}
-				/>
-			));
-		}
-		return <div>{orders}</div>;
+	const {onFetchOrders} = props;
+
+	useEffect(() => {
+		onFetchOrders(props.token, props.userId);
+	}, [onFetchOrders]);
+
+	let orders = <Spinner />;
+	if (!props.loading) {
+		orders = props.orders.map((order) => (
+			<Order
+				key={order.id}
+				ingredients={order.ingredients}
+				price={order.price}
+			/>
+		));
 	}
-}
+	return <div>{orders}</div>;
+};
 
 const mapStateToProps = (state) => {
 	return {
@@ -46,4 +47,4 @@ const mapDispatchToProps = (dispatch) => {
 export default connect(
 	mapStateToProps,
 	mapDispatchToProps
-)(withErrorHandler(Orders, axios));
+)(withErrorHandler(orders, axios));
